[#1829] - Agrupa las colecciones crudas en archivos por slug
Estaban declaradas dentro del agregador, con el argumento de que se
derivan del canon en vez de ser literales curados. No alcanza: el archivo
por slug es donde se busca una pieza del corpus, la escriba quien la
escriba, y las obras ya lo hacen así. El agregador queda solo
consolidando, que es su trabajo.

Entran en una carpeta propia, `onoff/collection/`, junto con su prosa en
Markdown: es hacia donde va a organizarse el resto del corpus, y crearlas
sueltas sería crearlas para moverlas enseguida.

La proyección que recorta la obra al shape de la query vive aparte, porque
la comparten las dos colecciones. Busca cada obra por slug y lanza si el
canon no la tiene, así una colección mal curada falla al importarse y no
en un caso suelto.

El glob de la regla que prohíbe importar una pieza puntual del corpus pasa
a cubrir subcarpetas: matcheaba un solo nivel, así que con las piezas un
nivel más abajo habría dejado de proteger nada, en silencio. Verificado
que sigue disparando desde la carpeta nueva.
