+++ src/mocks/onoff/README.md
@@ -47,7 +47,9 @@ La biografía de François Onoff vive como Markdown plano en un único archivo, 
 
 Corpus mínimo de dos colecciones de `LiteraryWork`, una por cada rama de `imagery`.
 
-- **Descripciones:** Markdown plano por colección — `geometrias-del-desvelo.collection.md` e `inventario-de-las-pasiones.collection.md`, importados con `?raw` y saneados con `markdownToSanitizedHtml`, misma convención que `<slug>.editorial-note.md` de `LiteraryWork`.
+Sus piezas viven agrupadas en `collection/`, no sueltas en esta carpeta: es la organización a la que van a migrar también las demás entidades.
+
+- **Descripciones:** Markdown plano por colección — `collection/<slug>.collection.md`, importados con `?raw` y saneados con `markdownToSanitizedHtml`, misma convención que `<slug>.editorial-note.md` de `LiteraryWork`.
 - **Colecciones:** `../onoff-collections.mock.ts`, export `geometriasDelDesveloCollectionMock` (rama `representative`, con portada editorial propia) e `inventarioDeLasPasionesCollectionMock` (rama `sample`, sin portada propia) — ambas construidas vía `createCollection`.
 - **Obras:** cada colección se cura con las obras que su propia prosa nombra —`geometria`/`losPeldanos`/`lasEscaleras` y `elTratadoDeLosPlaceres`/`elOdio`/`lasDosAntorchas`—, importadas del agregador por nombre. No cortar el agregador por índice: las dos colecciones quedarían indistinguibles por contenido.
 - **Agregador:** `onoffCollectionsMock: Collection[]`.
@@ -61,7 +63,7 @@ Contraparte cruda del corpus de dominio `Story` — lo que devuelven las queries
 
 - **Story raw:** `<slug>.raw.mock.ts`, export `<slugCamelCase>RawStory: NonNullable<StoryBySlugQueryResult>`.
 - **Storylists raw:** `<slug>.storylist.raw.mock.ts` (p. ej. `geometrias-del-desvelo`).
-- **Colecciones raw:** no tienen archivo por slug — se derivan del canon crudo de obras en `../onoff-raw-collections.mock.ts`, junto con sus teasers y sus escenarios de borde.
+- **Colecciones raw:** `collection/<slug>.collection.raw.mock.ts`, una por archivo. Sus obras no se escriben: las proyecta del canon crudo `collection/raw-collection.projection.ts`, que las busca por slug y falla al importarse si alguna no existe. El agregador `../onoff-raw-collections.mock.ts` las consolida y deriva de ahí los teasers, los selectores por capacidad y los escenarios de borde.
 - **Agregadores:** `../onoff-raw-stories.mock.ts` (`onoffRawStoriesMock`, teasers `<slugCamelCase>RawTeaser`, `onoffRawTeasersMock`, `onoffRawNavTeasersMock`); `../onoff-raw-author.mock.ts` (`rawOnoffAuthor`, `rawOnoffAuthorTeaser`).
 
 ## Corpus raw: `LiteraryWork` (#1981)
