+++ src/mocks/onoff/collection/inventario-de-las-pasiones.collection.md
